# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Service",
        "description": "Order API Documentation",
        "contact": {
            "email": "bugrabozkurtt@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "description": "Retrieves a list of all orders",
                "operationId": "e32d5b19d3160c49d1c56082d64be5c5",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Creates a new order and returns the created order details",
                "operationId": "41bf19b47f10a7a1a525e4a81646c87a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreatedOrderResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{orderId}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "description": "Deletes an order by its ID",
                "operationId": "e71770e644392db5eba60ca699278d11",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{orderId}/discounts": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get discounts for an order",
                "description": "Retrieves all discounts applied to an order",
                "operationId": "9bd1e50e72eba6ee9784e65d350a9f28",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order ID to get discounts for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDiscountsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/revenue": {
            "post": {
                "tags": [
                    "Users Revenue"
                ],
                "summary": "Get revenue for users",
                "description": "Retrieves revenue information for multiple users",
                "operationId": "fda1216f5798fbf19ae22a7dd5f7b6c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRevenueRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Revenue retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "revenue": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CartItemDTO": {
                "title": "Cart Item DTO",
                "description": "Represents a single item in the cart",
                "required": [
                    "product_id",
                    "unit_price",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "description": "ID of the product",
                        "type": "integer"
                    },
                    "unit_price": {
                        "description": "Unit price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "quantity": {
                        "description": "Quantity of the product",
                        "type": "integer"
                    },
                    "line_total": {
                        "description": "Total price of the item",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SyncProductStockDTO": {
                "title": "Sync Product Stock Data Transfer Object",
                "description": "Data structure for synchronizing product stock",
                "properties": {
                    "product_id": {
                        "description": "Unique identifier of the product.",
                        "type": "integer",
                        "example": 101
                    },
                    "quantity": {
                        "description": "Updated stock quantity of the product.",
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "CreateOrderRequest": {
                "title": "Create Order Request",
                "description": "Request payload for creating a new order",
                "required": [
                    "items"
                ],
                "properties": {
                    "items": {
                        "description": "List of items in the order",
                        "type": "array",
                        "items": {
                            "properties": {
                                "product_id": {
                                    "type": "integer",
                                    "example": 101
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "unit_price": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 99.99
                                }
                            },
                            "type": "object"
                        },
                        "example": [
                            {
                                "product_id": 101,
                                "quantity": 1,
                                "unit_price": 120.75
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "UserRevenueRequest": {
                "title": "User Revenue Request",
                "description": "Request payload for retrieving revenue of multiple users",
                "required": [
                    "user_ids"
                ],
                "properties": {
                    "user_ids": {
                        "description": "List of user IDs",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
            "CreatedOrderResource": {
                "title": "Created Order Resource",
                "description": "Resource representation of a created order",
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total amount of the order",
                        "type": "number",
                        "format": "float"
                    },
                    "discounted_total": {
                        "description": "Total amount after discounts",
                        "type": "number",
                        "format": "float"
                    },
                    "success": {
                        "description": "Indicates if the order was successfully created",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "OrderDiscountItemResource": {
                "title": "Order Discount Item Resource",
                "description": "Details of a discount applied to an order",
                "properties": {
                    "discountReason": {
                        "description": "Reason for the discount",
                        "type": "string"
                    },
                    "discountAmount": {
                        "description": "Amount of discount applied",
                        "type": "number",
                        "format": "float"
                    },
                    "subtotal": {
                        "description": "Subtotal before discount",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "OrderDiscountsResource": {
                "title": "Order Discounts Resource",
                "description": "Resource representation of an order's discounts",
                "properties": {
                    "orderId": {
                        "description": "ID of the order",
                        "type": "integer"
                    },
                    "discounts": {
                        "description": "List of applied discounts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderDiscountItemResource"
                        }
                    },
                    "totalDiscount": {
                        "description": "Total discount amount",
                        "type": "number",
                        "format": "float"
                    },
                    "discountedTotal": {
                        "description": "Total amount after discount",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "OrderItemResource": {
                "title": "Order Item Resource",
                "description": "Details of an item in an order",
                "properties": {
                    "productId": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Quantity of the product",
                        "type": "integer"
                    },
                    "unitPrice": {
                        "description": "Unit price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "description": "Total price of the item",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "OrderResource": {
                "title": "Order Resource",
                "description": "Order resource representation",
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer"
                    },
                    "customerId": {
                        "description": "Customer ID",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total amount",
                        "type": "number",
                        "format": "float"
                    },
                    "items": {
                        "description": "List of order items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Users Revenue",
            "description": "Users Revenue"
        }
    ]
}